корзина вынесена в отдельный компонент
